SIP-9649 - BE: Do not allow to publish or delete analysis from system category folder

+++ sip-services/sip-metadata-service/src/main/java/com/synchronoss/saw/analysis/controller/AnalysisController.java
@@ -1,7 +1,6 @@
 package com.synchronoss.saw.analysis.controller;
 
 import static com.synchronoss.saw.util.SipMetadataUtils.getTicket;
-import static com.synchronoss.saw.util.SipMetadataUtils.getToken;
 import static com.synchronoss.saw.util.SipMetadataUtils.validateTicket;
 import static com.synchronoss.sip.utils.SipCommonUtils.authValidation;
 import static com.synchronoss.sip.utils.SipCommonUtils.checkForPrivateCategory;
@@ -12,9 +11,9 @@ import com.synchronoss.bda.sip.jwt.token.Ticket;
 import com.synchronoss.saw.analysis.modal.Analysis;
 import com.synchronoss.saw.analysis.response.AnalysisResponse;
 import com.synchronoss.saw.analysis.service.AnalysisService;
-import com.synchronoss.saw.exceptions.SipAuthorizationException;
 import com.synchronoss.saw.util.SipMetadataUtils;
 import com.synchronoss.sip.utils.Privileges.PrivilegeNames;
+import com.synchronoss.sip.utils.SipCommonUtils;
 import io.swagger.annotations.ApiOperation;
 import io.swagger.annotations.ApiParam;
 import io.swagger.annotations.ApiResponse;
@@ -214,6 +213,14 @@ public class AnalysisController {
       analysisResponse.setMessage(HttpStatus.UNAUTHORIZED.getReasonPhrase());
       return analysisResponse;
     }
+
+    Long categoryId = analysis.getCategory() != null ? Long.valueOf(analysis.getCategory()) : 0l;
+    if (response != null && response.getStatus() != HttpStatus.UNAUTHORIZED.value()
+        && SipCommonUtils.haveSystemCategory(authTicket.getProducts(), categoryId)) {
+      analysisResponse.setMessage(HttpStatus.UNAUTHORIZED.getReasonPhrase());
+      return analysisResponse;
+    }
+
     analysisService.deleteAnalysis(id, authTicket);
     analysisResponse.setMessage("Analysis deleted successfully");
     analysisResponse.setAnalysisId(id);
